refactor: 深さの型をu32からu8に変更し、関連する部分を更新 fix: BestmoveEmitterの送信エラーハンドリングを改善

+++ packages/rust-core/crates/engine-cli/src/engine_adapter/utils.rs
@@ -227,7 +227,7 @@ impl EngineAdapter {
         Ok(ExtendedSearchResult {
             best_move: move_to_usi(&best_move),
             ponder_move,
-            depth: result.stats.depth as u32,
+            depth: result.stats.depth,
             score: result.score,
             pv,
             node_type: result.node_type,
